+++ CHANGELOG.md
@@ -9,6 +9,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ### Added
 
+- Add `:returning` option to `use Drill`
+- Allow user to set `:source` to table name instead of schema
 - Add `:repo` to `Drill.Context`
 - Allow user to override location of seeder file through `seeds_path` option when running `mix drill`. E.g. `mix drill -r MyApp.Repo --seeds-path priv/seeds/core`
 - Allow manual seeds and set callback `factory/0` to optional